Click and Refresh Event handling improvements
The click event on the select all checkbox was re-added on every
refresh, causing the select all event to run multiple times after
multiple refreshes. (Move the subscription to the same spot as where the
refresh is created).

Removed the subscription on the reference, this causes redundant
changes. Since the widget also catches its own change it will end up
doing a lot of unnecessary changes and refreshes.